+++ module/ability-dialog.js
@@ -25,20 +25,46 @@ export class AbilityDialog extends Dialog {
         });
     }
 
-    async _getContent(role) {
-        if (!role) role = "Spy";
-
+    static async getRollList() {
         let sourceCompendium = game.settings.get(
             "foundryvtt-quest",
             "abilityCompendium"
         );
 
-        //console.log(sourceCompendium);
+        let AllAbilities = [];
 
-        const QUESTAbilities = await game.packs.get(sourceCompendium);
-        //console.log(QUESTAbilities);
+        for (let i = 0; i < game.quest.AbilitySources.length; i++) {
+            //console.log(game.quest.AbilitySources[i]);
+            let QUESTAbilities = await game.packs.get(
+                game.quest.AbilitySources[i]
+            );
+            let compendiumAbilities = await QUESTAbilities.getDocuments();
+            AllAbilities = [].concat(AllAbilities, compendiumAbilities);
+        }
 
-        let AllAbilities = await QUESTAbilities.getDocuments();
+        //const QUESTAbilities = await game.packs.get(sourceCompendium);
+        //AllAbilities = await QUESTAbilities.getDocuments();
+        const roleList = [
+            ...new Set(AllAbilities.map((data) => data.data.data.role))
+        ];
+        return roleList;
+    }
+
+    async _getContent(role) {
+        if (!role) role = "Spy";
+
+        let AllAbilities = [];
+
+        for (let i = 0; i < game.quest.AbilitySources.length; i++) {
+            let QUESTAbilities = await game.packs.get(
+                game.quest.AbilitySources[i]
+            );
+            let compendiumAbilities = await QUESTAbilities.getDocuments();
+            AllAbilities = [].concat(AllAbilities, compendiumAbilities);
+        }
+
+        //const QUESTAbilities = await game.packs.get(sourceCompendium);
+        //AllAbilities = await QUESTAbilities.getDocuments();
         const roleList = [
             ...new Set(AllAbilities.map((data) => data.data.data.role))
         ];
@@ -138,3 +164,26 @@ export class AbilityDialog extends Dialog {
         html.find(".paths").click(this._scrollTo.bind(this));
     }
 }
+
+Hooks.on("renderCompendiumDirectory", (sidebar, html, _) => {
+    //console.log(html);
+    let addhtml = `<footer class="compendium-footer">
+    <span class="document-type">Loaded in Ability Browser</span>
+</footer>`;
+    for (let i = 0; i < game.quest.AbilitySources.length; i++) {
+        console.log(game.quest.AbilitySources[i]);
+        let element = html.find(
+            `[data-pack='${game.quest.AbilitySources[i]}']`
+        );
+        element.append(addhtml);
+    }
+    //let test = html.find(`[data-pack='${current}']`);
+});
+
+Hooks.on("getCompendiumDirectoryEntryContext", (html, entryOptions) => {
+    entryOptions.push({
+        name: "Toggle Ability Browser",
+        icon: '<i class="fas fa-edit"></i>',
+        callback: (e) => game.quest.api.toggle(e[0].attributes[1].nodeValue)
+    });
+});