+++ changelog.md
@@ -1,5 +1,10 @@
 # Quest Foundry VTT System
 
+0.1.11
+
+-   Added API hook event to register compendium to the ability browser
+-   Added ability to add compendium to the ability browser through the right click context window of the compendium
+
 0.1.10
 
 -   Updated Combat Tracker to have combat groups and allowing rolls from the tracker.